Hammer of Light - Energy cost: 30 -> 50 - - Mechanics: Current ->
Throw down a Hammer of Light on the ground, dealing 196-304 damage every second to nearby enemies and healing nearby allies for 154-226 every second in a 6 block radius. Your Protector's Strike pierces shields and defenses of enemies standing on top of the Hammer of Light. Lasts 10 seconds.

You may SNEAK to 'pick up' the Hammer of Light and turn into a temporary Crown of Light to carry the effects of hammer with you in a 6 block radius. However, it removes the damage and piercing BUT increases hammer healing by 75% and reduces the energy cost of your protector strikes by 10 energy. You cannot put the Hammer of Light back down after you converted it.

The crown will have the same duration as hammer, so quick decision making is advised. Aka, you cannot make both phases last 10 seconds if you use it last second.