**Revenant**
Reckless Charge
- Damage: 466-612 -> 457-601 (-2%)
- Cooldown: 9.98s -> 9.32s

Orbs of Life
- Orb Healing: 240-350 -> 250-375

Undying Army
- Flat healing: 175 -> 200

**Protector**

Hammer of Light
 - NEW:

Energy cost: 50
Cooldown: 62.64s

Throw down a Hammer of Light on the ground, dealing 178-244 damage every second to nearby enemies and healing nearby allies for 178-244 every second in a 6 block radius. Your Protector's Strike pierces shields and defenses of enemies standing on top of the Hammer of Light. Lasts 10 seconds.

You may SNEAK to 'pick up' the Hammer of Light and turn into a temporary Crown of Light to carry the effects of hammer with you in a 6 block radius. However, it removes the damage and piercing BUT increases hammer healing by 60% and reduces the energy cost of your protector strikes by 10 energy. You cannot put the Hammer of Light back down after you converted it.

**Aquamancer**

Healing Rain
- Healing: 190-250 -> 200-250
- Crit Multiplier: 175% -> 200%